+++ backend/store_api/Dtos/Discounts/DiscountUpdateDto.cs
@@ -5,11 +5,11 @@ namespace store_api.Dtos.Discounts;
 public class DiscountUpdateDto : IBaseDto<DiscountEntity>
 {
     public Guid? ProductId { get; set; }
-    public double? Percentage { get; set; }
+    public int? Percentage { get; set; }
     public DateTime? StartDate { get; set; }
     public DateTime? EndDate { get; set; }
 
-    public DiscountUpdateDto(Guid? productId, double? percentage, DateTime? startDate, DateTime? endDate)
+    public DiscountUpdateDto(Guid? productId, int? percentage, DateTime? startDate, DateTime? endDate)
     {
         ProductId = productId;
         Percentage = percentage;
