+++ backend/store_api/Entities/DiscountEntity.cs
@@ -30,11 +30,11 @@ public class DiscountEntity
 {
     public Guid Id { get; set; }
     public Guid ProductId { get; set; }
-    public double Percentage { get; set; }
+    public int Percentage { get; set; }
     public DateTime StartDate { get; set; }
     public DateTime EndDate { get; set; }
 
-    public DiscountEntity(Guid Id, Guid ProductId, double Percentage, DateTime StartDate, DateTime EndDate)
+    public DiscountEntity(Guid Id, Guid ProductId, int Percentage, DateTime StartDate, DateTime EndDate)
     {
         this.Id = Id;
         this.ProductId = ProductId;
